requirements.txt
Django==5.2.2
django-ninja==1.4.3
shortuuid==1.0.13
django-ninja-extra==0.30.1
django-ninja-jwt==5.3.7
django-tailwind==4.0.1
django-browser-reload
pytest-cov==6.1.1
pytest-django==4.11.1
psycopg2-binary==2.9.10
ruff # for format